limit the size of images

+++ v1/services/image_service.py
@@ -8,32 +8,41 @@ from v1.models import Card
 
 
 class ImageService:
-    def __init__(self, max_dim: tuple[int, int] = (800, 600), image_format: str = 'PNG'):
-        self.max_dim = max_dim
-        self.format = image_format
+    def __init__(self, max_size_mb: float = 1.0):
+        self.max_size = self._megabytes_to_bytes(max_size_mb)
 
     def save_image(self, card: Card, image_file: UploadedFile):
+        if len(image_file.read()) > self.max_size:
+            return HttpResponse(status=HTTPStatus.REQUEST_ENTITY_TOO_LARGE)
+
         if card.image:
             card.image.delete()
 
-        image_file.name = f"{card.id}.{self.format.lower()}"
+        image_file.name = f"{card.id}__{image_file.name}"
         card.image = image_file
         card.save()
 
         with card.image.open() as image:
-            return self._create_response(card, image.read())
+            return self._create_response(image)
 
     def get_image(self, card: Card):
         if not card.image:
             return HttpResponse(status=HTTPStatus.NO_CONTENT)
 
         with card.image.open() as image:
-            return self._create_response(card, image.read())
+            return self._create_response(image)
 
-    def _create_response(self, card: Card, image: UploadedFile):
+    @staticmethod
+    def _create_response(image: UploadedFile):
+        extension = image.name.split('.')[-1]
+
         response = HttpResponse(
-            content_type=f"image/{self.format.lower()}",
-            headers={"Content-Disposition": f'attachment; filename="{card.id}.{self.format.lower()}"'},
+            content_type=f"image/{extension}",
+            headers={"Content-Disposition": f'attachment'}
         )
-        response.write(image)
+        response.write(image.read())
         return response
+
+    @staticmethod
+    def _megabytes_to_bytes(megabytes: float) -> int:
+        return int(megabytes * 1024 * 1024)
